requirements.txt
fast_api==0.115.6
python-dotenv==1.1.0
groq==0.22.0
sentence-transformers==
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.6.1